moved mailchian cmd handlers
