+++ cmd/mailchain/internal/rest/handlers/read.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package read
+package handlers
 
 import (
 	"encoding/json"
