+++ cmd/mailchain/internal/rest/handlers/send_message_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package send
+package handlers
 
 import (
 	"encoding/hex"
